Add intellihide status var preventing unnecessary show/hide call.
Currently a check before the execution of show/hide functions is implemented
inside the dockedDash object, relying on the autohide variable since
currently show/hide function are linked to the activation/deactivation of
the autohide effect. As I'm planning to change or at least implement as an
option also different behaviours, this will be useful in the future. Moreover,
this is the place where it makes more sense that such check is performed.