Try to make the extension more theme undependent.
I create another actor with name #dash. This serves for applying an opaque
background for those themes like the default one that has a semi-transparent
dash. I inherit all dash style of the current theme, then disable all those
non interesting. I'm interested only on the shape, thus only on the border
radius I think, in order to cover all and only the dash area. It is probably
a little ugly workaround, but I have not found a way to access the current
style and simply change the background alpha.