Remove non existent tweener owerwrite property.
In the current code overwrite was not necessary, as the default always
owerwrite mode is enough. However, I discovered that in gjs tweener version
(1.31.71) that property is not present. The online documentation liked in gjs
home page refers to a new version and it is a little deceiving in this regard.