Correctly set initial Y position and visibility.
I removed the animation of the dash Y position with a static setting of it.
This because now I think the correct y position is determined when extension is
loaded, and because the animation if triggered erlier was causing problems.
However a situation when the animation would be desiderable is when theme is
changed. This will be taken into account later.