Enable dash accessibility via ctrAltTab.
Accessibility via ctrlAltTab was no more working with the extension enabled.

I had to take into account that the dash is often hidden and has to return
hidden! So Dash is shown when focus is selected via ctrAltTab interface. Dash
is hidden by mouse hover motion, pressing Esc or when it loses focus.

This is a port of an equivalent commit in the gnome-3.4 branch.