Improve autohide when popup menus are displayed.
A timeout loop is started when dock is shown forcing sync_hover when popup
menus are close. Actually sync_hover is called whenever popup menus are not
showing. Popup status is retrieved by looping over all dash app icons. I havn't
found a more direct signal of a popupMenu being closed.

This is a port of an equivalent commit in the gnome-3.4 branch.