Option: dash vertical centering and height increase.
Add a container actor with affectsInputRegion disable. The child actor is the
activated by manuallly calling chrome._trackActor().  Add schema keys and
pref.js settings.

Side effects: before the dock position was obtained from the staticBox, now the
opposite is done; Some actor were renamed.

This is a port of an equivalent commit in the gnome-3.4 branch.