Implement application based intellihide.
Optionally only windows of the current focused appplication are taken into
account by intellihide when this option is enabled. Also half maximized windows
are considered so that it's possible to work with two windows side by side.

This is a port of an equivalent commit in the gnome-3.4 branch.

Conflicts:
	intellihide.js