Add support for rtl languages.
By default the dock is placed on the right side of the screen following the
overview mirrored layout.

Set the dock anchor point to NORTH_EAST in rtl-mode so the dock doesn't move
when its width changes. Update _set_clip() and _updateStaticBox() accordingly.

This is a port of an equivalent commit in the gnome-3.4 branch.