App based intellihide: better focused app identification.
If there is not a focused window, consider the window that is on top of the
others of the current workspace to find the currently focused application. This
happenis when focus is given to the desktop window, e.g. clicking on the
nautilus desktop window, when focus is given to some chrome actors such as a
notification or when upon closing the focused window the intellihide check
function is triggered before the next one get focus.