Bug: move anchor points only at the  animation start.
Previously the anchor point was moved from EAST to WEST and viceversa right
when calling _animateIn() and _animateOut(). If another animation was running
and the new one  was added with a delay the old animation loop kept updating
the actor position to final_position based on the worng old anchor point.