Minor changes and improvments to the dock actor containers.
Set affects struts to the dash._box via trackChrome instead of to the
container, so that the 1px wide padding isn't included in the struts. The main
actor is not removed and readded anymore on settings change. Input region is
still applied to the container _box, since I need to capture mouse events also
in the padding area.