Improve application button behaviour.
Keep the extension application button and the default one in sync. Let it do
all the job apart from entering and exiting the overview when needed.

Also in the previous commit the application button inside the overview was
replace by the extension one and not correctly resetted when the extension is
disabled.