Integrate the DropDownTerminal extension
The DropDownTerminal extension window hint is POPUP_MENU and we do not support
iit yet. Moreoveri, the terminal windows is by design not a proper application
and this breaks the application based intellihide. Thus the window is
whitelisted by its wm class.

This fixes
https://github.com/zzrough/gs-extensions-drop-down-terminal/issues/24 on the
Dash to Dock side as it is way simpler to do that than the other way around.