Do not exit from accessibility automatically dropping basic keypress handling.
Following the upstream changes, there is now a windows group to focus back on
the windows, thus remove the keypress handling and all now unused helping
functions.

This is an accessibility function while when I wrote this code I had in mind
the idea of adding keyboard shortcuts. I will add back again similar code if I
will consider to add keyboard shortcuts. Note that the  isMenuUp variables I
was monitoring to keep the dash visible was removed from upstream and a new
trick will have to be found.