extend-height is "real" 100% of monitor height
Panel is moved to right (or left when rtl) when
extend-height and dock-fixed options are selected.

This works nice with a plugin that (auto)hides
main panel.

Known bug: after this extension is enabled, main panel
x position is set to 100, which is not the width of the dash.
this.emit('box-changed') has to be called for this to be fixed.