Handle smooth scrolling when switching workspace (solve bug)
Starting from gnome-shell 3.8 smooth scroll events are supported. Instead of an
UP or DOWN event a SMOOTH event is triggered sometimes when scrolling over the
dash. Since the  SMOOTH case was not handled by the switch case, a deadtime was
set but the workspace was not switched when using the
scroll-switch-workspace-one-at-a-time option.

Handle the SMOOTH case corectly. Always check if a workspace is actually moved
before setting a new deadtime.