Solve animations bug: dash gets stuck with very frequent hover on/off.
I was avoiding to go through the tweener animation if no animation was supposed
to be meaingfull. Previously If the final position was already the actual
position, the tweener was not called at all to avoid unnecessary computations.
However, this results in the animation status to become inconsistent with very
frequenct hover on and off events (Mmmmh, still not clear why... I thought to
have understood, but it doeasn't make sense again now!). Always initializing
the tweener seems to prevent the problem.