Bug: workaround to prevent allocation cycle.
A proper implementation would probably require to write a layout manager or
write the parent actor allocation functions. Not sure why there is the
allocation cycle in the first place. Upon allocation of actor 'b' I change
actor 'c'. both 'b' and 'c' are children (not directly) of a common parent 'a':

a > b, c

This is just a workaround of the problem.