Bug: the dash object was never destroyed.
The dash object contains globals signals that are disconnected by the destroy
method. However, while I was destroyng the dash main actor and thus all its
children, the destroy method was never called. An alternative would have been
connect the destroy method to the destroyed signal of the main actor.