Solves bug caused by wrong hover status reported on the dash actor.
Close issue #124.

It seems that in Clutter the mouse pointer remains linked to the previous actor
when moving to a non reactive one. The change-hover is triggered and the hover
updated but, calling later manually sync_hover on the actor causes the hover
status to be wrongly calculated. No hover-change sygnal is triggered actually.

This bug was introduced by making the workspace switcher non reactive
(880e8941f0daede4e251c1449a0863371387918a). Just remove the code calling
sync_hover and forcing a recheck of the hover status as not necesary anymore.