Use #dashtodockContainer:overview pseudoclass instead of #dash:desktop
I changed my mind. The argument of atom-dock - don't break style when extension
is disabled - doesn't apply to dashtodock. I prefer to follow upstream choice
where for instance the overview pseudoclass is applied to the panel. I move the
pseudoclass to the outer #dashtodockContainer for consinstency and to allow
full customization.

Update built-in style accordingly.