Hide dashSpacer when not needed. Fix missing appIcon sping animation.
As reported in #136, appIcon spring animation was not played in the case
of 1 only static workspace, as no allocation of the appGrid is executed
given that neither the default dash - or better the default dashSpacer
which is bind to the dashSlider size - nor the workspacesThumnails are
visible whose sliding out triggers the animation.

Following the upstream behavior, hide the dashSpacer when entering the
search view. This has the effect of triggering the animation.