Solve broken appsGrid showing by SUPER+A shortcut.
Previous commit broke the showing of the appsGrid directly by SUPER+A as
it was dealing with the grid opacity which was not restored. Use the
hide method instead. As a consequence the grid needed to be hidden on showing.

Moreover, wait for the overview to be shown before animating in the
grid, as sometime wrong final positions were calculated.