myDash: after previous two commits (revert) make horizontal rtl works.
The bindConstrain ensure the DashActor is allocated a size equal to the
maxSize, thus forcing the showAppsIcon to be always shown and the
applications to be cropped if not enough space is available.

This ensure it works again even in horizontal rtl mode, which was the
reason why I went for a manual settings of the dash size.