Use Shell.GenericContainer for the DashSlideContainer.
According to this

https://bugzilla.gnome.org/show_bug.cgi?id=688973

it's not a good idea to extend Gobjects in an extension. If the
extension get's reloaded at runtine - this as far as I know happens only
whne  updating the extnesion from the extension website - those Gobjects
get's redefined with the same name, which is not allowed. This results
in a silent fail of the update process, which is quite annoying.

I revert to use the hack employed before virtual functions were availble
in Gjs.