Add timeout to check the hover on pressure sensed.
A drawback of not starting with an hover status when the dock is shwon
is that if the mouse goes away quick enough no hover change happen and
the leave event never get triggered. To avoid the dock from getting
stucked for an hover check after the animation out is finished.