myDash: Fix leaking signal in the appIcon popupMenu.
This was fixed upstream, but I dind't port in both places I use the
affected method. I connect to the destroy signal of the menu instead of
the actor, as the menu destroy follow the actor destroy and the callback
signal requires the existence of the menu, which could have been
destroied for other reasons.