Remove padding from the dock container.
The built-in theme set a 1px padding that made the launcher not
clickable in particular in screen edge side, which was inconvenient as
the mouse might actually rest there after triggering the dock showing.

Rather then removing it only on the screen edge side, removed it
everywhere as it doesn't seem necessary.