Intellhide heuristics: consider topmost windows in the current monitor.
The idea is that we dont want the dock to  overlap with the windows of
the topmost application, event is it's not the focused app, for instance
because in multimonitor the user select a window in the secondary
monitor.

This solves the situation in which a window is maximized in the primary
monitor and the dock is shjown upon selecting a window in the secondary
one. This also avoids showing and hiding the dock at each interaction
with the two monitors.

This might hide the dock more often than expected, when wotk effectively
happen on thre secondary one. However bein the dock from the working
area it doesn't harm too much to hide it.

This changes doesn't affect intellihide if the per-app option is not
selected.