Intellihide heuristics: always consider maximized windows.
Previously half-maximized windows were considered, so it makes sense to
consider fully-maximized windows, as it is reasonable to assume it is an
important window in the current workspace - like temporary working on
another window but still related to the main task. Otherwise the person
could have moved to another workspace.