Fix bug with top positioning in fixed mode.
Fix an infinite loop struts change-> dock position -> struts change with
the dock in fixed mode and top positioned.

The behaviour is now to always put the dock at the screen edge,
regardless of the presence of anything there, while avoiding elements
affecting struts in the other direction. The idea is that if you want
the dock at the top you most likely don't have the panel there.