Intellihide: revert to tracking windows size and position change.
Tack each window position and size change instead of relying on the grab
eventrs plus specific events (maximize, etc).

The reason for this is that not all situation in which a window change
position can be tracked, i.e. windows moved by the move action.

This appears to be less efficient, as way more events are triggered
while moving and resizing windows. To mitigate this a deadtime is
implemented to limit the loops over windows to check for overlap.