Intellihide revert to tracking directly windows changes.
I abandone the use of the grab-op-begin / grab-op-end as trigger for the
checking of overlap between the windows ans the dock, and move back to
tracking each individual windows size and positon change.

This have one main benefit: it is simpler because all windows changes
are tracked regardless of the origin of the change, including
keybindings, maximize/unmaximize etc., allowing to remove additional
watched events, implemented  workarounds, and residual edge-case bugs.
Overall it shoud be more robust.

The main drawback -- which was the motivation for discarding this method
-- is that way more events are potentially triggered, and with
uncontrolled high frequency. This will be optimized later.