Intellihide: fix glitch on windows creation.
When a new window is created it might be put at the top left corner of
the workarea, before being positioned according to the mutter windows
system. Although the window is never shown in such position, the
intellihide retrieve such position and start hiding the dock, which is
readily reshown as soon as the window position is updated. This result
in an annoying dock animation.

Checking for the mapped state seems to prevent this behaviour.