Track the struts of the dashtodokBox.
Commit 976bafaee28180a0102238589087d72817a78233 introduced a bug a long
time ago. The dash.actor is larger than the visible dash, as it is used
with a bindConstrain to set the dash max size. The result is that in
fixed mode the area below the dash become insensitive.

I moved to track this actor to avoid including the 1px padding once
required to show the dock. This is not a problem anymore (I'm using the
pressure barrier/dwelling sytem now), so reverting to tracking the
dashtodockBox fix this problem.