Fix regression with inputRegions in fullscreen and fixed mode.
Introduced by: 700a4d024ad0b4f1f563033e6505de28c1ca6673. The fullscreen
tracked actor gets hidden but the parent (_slider) whose input regions
are tracked is not.

The reason why it was working before is that I was traking the dash.box
acotr, which when hidden in fullscreen made  in turn the slider size
become 0. Tracking the the outer _box did not give the same result and
it is not clear to me why.

I move fullscreen tracking to the top actor.