Fix regression with inputRegions in fullscreen and fixed mode.
Introduced by: 700a4d024ad0b4f1f563033e6505de28c1ca6673. The actor which
is tracked for fullscreen has to be a parent of the one tracked for
input regions, otherwise the area remains sensitive to inputs blocking
them to reach the underlying window.

The reason why it was working before is that I was traking the dash
actor, which being hidden on fullscreen made the in turn the slider size
(whose input are tracket) of size 0. Tracking the the outer _box did not
give the same result and it is not clear to me why.