Track top level actor for fullscreen, fixing a bug.
This seems the opposite of what stated in my previous commit:

a1f63c706f34687d061f51f72d44256b71e87d06

I do not know what the truth is, but:

1. at the moment there are problems when fullscreening the windows of
certain applications in fixed mode (repeated showing/hiding of the
dock/toppanel), in partcular VirtualBox when the dock is at the botttom
of the screen, and occasionally also with Totem.

2. The "affectsInputRegion" was added back in 3.16, so it can be safely
used again. This allows me to track the toplevel actor.

This seems to solve the above mentioned problem.