Revert "Fix unnecessary return value."
This commit was completely wrong, I don't know what i was thinking
about. In the attempt of removing a warning for the return value not
always being specified, I ended up mapping to an empty array of window
all the times, resulting in the windows previewlist to be redrawn all
the time it was updated.

This fix issue #471.

This reverts commit bd2c3d7f88f86ab150db33a63e701c96e877805b.