Fix missing St import in convenience.js.
St is not imported in convenience.js. Although I never experience any
issue (St is defined globally by the shell itself?), it was noticed by a
person that thi was  causing the extension to not start with error:

Extension "dash-to-dock@micxgx.gmail.com" had error: ReferenceError: St
is not defined

It it not clear to me, but it is safe to import it and this fix the
issue for this person.