Tweak windows previews behaviour with custom click.
The current behaviour is to activate the application directly rather
than then swhowing the previews when only one windows is available. For
secondary/modified clicks it makes more sense to assume that one really
wants to trigger the previews menu rather than switching to the app.
Always show the previews in such cases.