Add 'toggled' signal to facilitate reconnecting to dock events.
With the introduction of the multi-monitor option, the dock is no longer
a single object but an array of objects. Add a 'toggled' signal to the
dockManager let other extensions know when the dock(s) have been
destroyed and recreated. This facilitates reconnecting to the dock(s)
events such as showing, hiding, and mouse hover.