Fix mouse trapped in monitor with fullscreen windows.
With autohide disabled in fullscreen and push to show enabled, when a
windows is in fullscreen the pressure barrier does not allow the mouse
to traverse the screen edges effectively trapping the mouse on one of
the monitor of a multi-monitor configuration. Normally (without
fullscreen windows or with the intellihide enabled in fullscreen),  we
remove the barrier just after showing the dock, releasing the mouse
movement.

Fix the issue by disabling the barrier with windows in fullscreen and
autohide in fullscreen disabled.