Monitor-isolation: move 'window-entered-monitor' from appIcons.js to docking.js.
On X, this wasn't necessary, as when migrating monitors a 'restacked'
signal was emitted anyway. In Wayland, we need this signal at a higher
level, as the appIcon instance might not even exist on the other dock.

This is similar to the addition of the 'switch-workspace' signal: only
really needed for Wayland.