Enable loading Dash to Dock alongside this extension
This extension will be virtually and dynamically disabled if Dash to Dock is
loaded and reenabled once Dash to Dock is disabled.
We also avoid loading on Shell start if Dash to Dock is in the extension
enabled list.

Basically, this allow the user to go into the tweak mode by just installing
that extension.
Note that some keybindings handler are not properly destroyed on main object
unload and can make the extension appears in error while still functional.
This is an upstream Dash to Dock issue that we should fix there.