Glossy launchers: set max computed color cache size.
We store the computed colors by applicatoins id in a global cache.
Compared to having cache lifetime linked to the appicon lifetime, this
has the advantage of avoiding to recompute it in the case:
1. We have more than one dock showing the same application.
2. An application is stopped/relaunched

However at the moment the cache could grow indefinitely. Even though it
is unlikely that more than a reasonable number of applications are run,
set a maximum number of stored items, with the expectation that it will
never be actually reached in practice.

We also never invalidate the cache, for instance for the case the
application icon changes, but this doesn't seems to me a big issue.