Monitor isolation: limit the impact of connecting to 'window-entered-monitor'.
In Wayland sessions, this signal is needed to track the state of windows
dragged from one monitor to another. However, the performance impact is
questionable as it triggers a redraw of all icons' style quite often.
To limit this, we only really connect to this signal if
'isolate-monitors' is true, and if there are at least 2 monitors.