Rework extended ShowAppsIcon machinery.
Background: we cannot extend a GObjects directly becuase they are static
objects, while the extension is dynamically loaded/unloaded.  We
previously live extended the instance. However, this results in a warning
in 3.26: "WARNING: addSignalMethods is replacing existing ..." due to :

https://bugzilla.gnome.org/show_bug.cgi?id=619710

as we were registering the signals twice. To overcome this, we now
wrapper the showAppsIcon class within a new class, and do the appropriate
rewiring upon its use.

Bonus: the end result is also slightly more elegant.