Load gettext dashtodock domain before calling gettext()
appIcons.js is assigning const __ = Gettext.gettext.
This is what enables to have "All windows" and other non Shell strings
(dash to dock specific) translated.
However, the domain was set in init(), after this assignement, which was
thus loading the GNOME Shell domain. Ensure we bind the domain before
(which is a really fast call).