Update dockManager variable declaration for Gnome 3.26 GJS (mozjs52).
dockManager properties were not accessible by other extensions (e.g.
workspaces-to-dock) after locking/unlocking the Gnome 3.26 desktop or
after disabling/enabling the extension. What's weird is that it seems to
work fine on the initial login to the desktop. Changing the dockManager
declaration from 'let' to 'var' fixes the issue.