appIcons: Implement window tracking for removable devices and trash
In Unity, special logic is present that will map a Nautilus window
the removable device or trash icon in the dock. The key to making
this possible is a special dbus property that was patched into
Nautilus that allows us to find the window where a location is
open.

Once we have access to this Nautilus information, we can then jump
through a bunch of hoops to map the locations to MetaWindows and
then a little special-casing logic, link our dock icons to those
windows.

Now, the special icons will have a running process highlight and
window counts, and all the usual features of a running app.

We support both the Ubuntu-specific patched Xid based window
matching as well as the upstream GtkApplication based matching
that I added to Nautilus.

In a difference from Unity, I made no attempt to subtract the
special location windows from the Nautilus app; that would be a
bunch of work and the benefit is unclear.

When run with a Nautilus that supports neither method, we will
simply never see any linked windows, and the behaviour will be
the same as without this change.