dash: Add Trash Icon
This change introduces a basic trash icon that can open and empty
the trash. I did not attempt to implement any sort of drag and
drop to put something in the trash.

The implementation follows the path of least resistence and creates
a DesktopAppInfo so that it can be represented by a regular AppIcon.

That has particular implications - most significantly being that
any actions offered by the icon have to be Execs of external programs,
rather than code.

In this case, we care about opening the trash and emptying it. We
can use the 'gio' utility to open the trash using the default file
manager, but emptying it is trickier.

You can use gio to empty the trash but there is no confirmation
dialog if you do this. Rather than implement such a dialog, I
decided to use the Nautilus EmptyTrash dbus call; this triggers
confirmation from Nautilus, but is obviously nautilus specific.

Finally, I did not attempt to pin the icon to the bottom of the
dash as Unity does. Given how elaborate the icon allocation logic
is, I couldn't bring myself to tackle it.