dash: Add Removable drive/device icons
Another Unity dock capability is showing icons for removable drives
and devices.

This change introduces such icons for these entities. As with
the Trash, we back these with DesktopAppInfo, and implement the
open/unmount/eject actions with the 'gio' utility.

In Unity, icons are shown for both mounted and unmounted entities,
and this behaviour is retained. Also retained is the practice of
not exposing an unmount operation for ejectable entities. We also
cannot show an unmounted icon if the entity has no activation root,
but I believe that most entities of this type are ejectable so it's
not a real problem. This limitation arises because the activation
root is how we know where to mount the entity.

Unlike the trash, the natural icon placement matches the behaviour
in Unity.