Enhance adaptive mode avoiding perceived panel flickering
When the Dock is hidden (autohide/intellihide), and a window is next to
it without touching the panel, the panel was going opaque. Disable the
computation of the  panel state when in adaptive and in
autohide/intellihide. This ensures that the dock itself has the correct
transparency/opaque state when revealed above a window (independently of
the panel), i.e. the panel opaque state also makes the dock opaque but
not viceversa.  This avoid a perceived panel flickering (Fixes #636).