Enhance adaptive mode avoiding perceived panel flickering
When the Dock is hidden (autohide/intellihide), and a window is next to it
without touching the panel, the panel was going opaque.
Disable the dock computation for panel state when in adaptive and
in autohiding mode. This ensure that the dock itself though has the correct
transparency/opaque state when revealed above a window (independently of
the panel).
Fixes #636.