AppIconIndicators: correctly update drawing_area.
st_drawing_area_queue_repaint seems to be the right method instead of
clutter_actor_queue_redraw, that according to the documentation only
causes the same content (without style update) to be redrawn.

There was a delay in the update of the indicator style, for instance
when changing the settings.