Fix some Gjs warnings.
Following upstream changes in Gjs (See
https://gitlab.gnome.org/GNOME/gjs/merge_requests/22, Gnome Shell 3.26+)
warnings are reported when properties of destroyed (finalized) GObjects
are accessed.  I can't seem to find a way to to identify such condition.

The appIndicator destroy code take care of destroying the actors created
and restoring the default styles. It is not necessary to do this when
the appIcon is destroyed. Thereforewe don't explicitely call the destroy
method when the appIcon is destroyed, instead just disconnect the global
signals.