Fix crash connecting/disconnecting monitors.
We were assuming the presence of at least one monitor (the primary one).
Instead, we should take into account the headless configuration which
can also temporary occurs while disconnecting and reconnecting monitors.

As we are destroying and recreating the docks whenever the monitor
configuration changes, stop trying to recreate the dock(s) if there are
no monitors connected.

We might have started to be subjected to this bug after upstream commit:
http://gitlab.gnome.org/GNOME/gnome-shell/commit/5f8a5114833d7

Fix #627.