Use a wrapper to access to Screen, Workspacess and Monitor
As per mutter xwayland-on-demand preparation branch refactory MetaScreen
has been removed, thus replacing the dead objects.

Basically global.screen moved to global.display, while all the workspace
related features have been moved to the WorkspaceManager and displays
are managed by DisplayManager.

In order to keep the extension working in both 3.28 and 3.30 gnome-shell
we need to support this change by creating wrapper functions that use
the gnome.screen if that is available or the proper new object
otherwise.