Fix for border radius with adaptive/dynamic opacity.
The transparency machinery is overwriting the inline style of the dock,
as is the code to adjust the border radius when the position of the dock
is not left.

Ensure we append the opaque/transparent style to any existing styles.
We move the computation of the transparency style in the transparency
enable stage, which is always called when restting the adjusted
border-radius styles.

Fix #666.