Fix mixing of es6 class names with St names.
The latter are used as ids to address the objects in the stylesheets,
while the former are meant to represent the function/class name as
declared.

This was resulting into a bug as the '#Dash' identifier was missing and
do default style (e.g. background) was taken from the stylesheet.