Don't reset border-radius for shrinked dash
In stylesheet.css, the border radius was set to 9px for the shrinked
and the built-in theme for dash, in order to reduce its size.

Since the default shell theme is already using 9px border-radius for
dash, this didn't really change anything. Additionally, when using
custom user-theme for gnome-shell, the border-radius defined in
dash-to-dock's stylesheet.css overrides any user-themed stylesheets,
possibly breaking the appearence of the shrinked dash with custom
gnome-shell themes.

This makes dash-to-dock use same border-radius for shrinked and
non-shrinked dash, to prevent issues with user-themes.