Fix mouse trapped due to barrier set up with dock shown.
We were creating a barrier after each call to the updateBarrier
callback, regardless of the current dock state. If the dock is already shown,
this creat a barrier which is never removed, resulting in the mouse
being trapped. Check for the current dock state instead.

This is triggered for instance when an application enter/exit
fullscreen on any monitor.

The logic controlling pressure barriers and dock state needs
improvements but ths should fix it for now.