Style focused backgrounds only for built-in theme
Focused application's icon was highlighted with light background
in stylesheet.css when using custom window counter indicators, or
dash-to-dock's built-in theme.

Since that highlight style isn't necessarily visible, or simply
doesn't look good with some gnome-shell themes, it seems better to
style the background only when using the built-in theme, and use
default focused styling from the shell theme otherwise. For that
reason, stop styling the focused icon when using custom indicators.