Fix regression with border radius with built-in theme.
This was introduced by f2ce64a6277945b3a5494a090466d5c955162d65.

Although it make sense for the shrinked case to just use the theme
defined styles, for the built-in theme the idea is that we do not
overwrite anything dynamically but just use statically defined style
rules. The reasoning is that this is meant to support the defualt theme
only, so we know what the style look like so it doesn't harm to tweak it
directly, and at the same time we are less prone to bugs in this way as
compared to the dynamic case which is instead a best effort atttempt to
support as many themes as possible.