Fix bug with xscreensaver.
The dock is visible when xscreensaver is triggered. This occur in Gnome
Shell 3.34, but not in 3.30. Slighly changing the position of the dock
in the stack seems to fix the problem. This might be a consequence of
upstream commit 2bd80579eddba9804871ce6c8de742bdf9b1d800.  This also
requires to disable the machinery in place to keep the dock on top of
everying on screen grab seems. Luckily, this seems to not to be
necessary anymore.

Fix #1002