docking: Delay hide requests if we're currently showing
Wait that the show animation is completed before we try to hide a sliding
dock.

This fixes a regression caused by commit 662604b because with clutter
animations, a transition would remove the previous one queued for an object,
and thus the delay wouldn't work anymore.

So, instead of relying on the waiting time, let's just use the _animateIn()
onComplete() callback in order to request an hide while the show animation
is still going.