docking: Ensure that all the slider children are properly centered
Not to make the whole dash container to stick to the topmost/leftmost side,
we need to adjust the vertical and horizontal alignment of the dash slide
container so that it's centered for the relevant axis.

This was previously applied to the whole Docked dash, but this was already
well positioned, while we need to ensure that its content is also properly
aligned.