dash: Reuse as much as possible upstream code
Many functions in dash were just copy-pasted from upstream source, and this
may lead to living with non-updated code, so instead of this just re-use
the default methods, given that JS is nice (or not?) enough to allow to do
that easily.

The only difference is in handleDragOver where we had to temporarily mock
some properties in order to adapt the code for the horizontal mode and to
compute the icons size based on the actual scrollview content, instead of
the visible box area.

Less code to maintain, yay!

LP: #1867763
LP: #1867613