locations: Offer to filter out devices by regex
People have asked for the ability to filter out certain devices from
appearing in the doc and today it is all-or-nothing.

This change is somewhat a proof-of-concept as the user configuration
is literally a regex, which is very powerful but not terribly
intuitive. I would be interested in suggesitons on how to handle the
configuration.

Fixes #1081