Fix positioning and icon size calculation.
The size constrain which is used by the machinery to calculate the dock
icon size seems to not work unless the Actor is assigned a FILL align.
It's not clear to me if this is a consequence of changes upstream
possibly related to the deprecation of the St.Align in favor af
Clutter.ActorAlign, which we applied in
66a17dc1aca9d606dfab376f5eb7b8d51eb880f9.

The problem is only occuring if the dock is not in extended mode.
These seem (a bit empirically) to restore the behaviour.