docking: Properly replace default dash in all modes
As per 3.36, the dash is accessible via Main.overview.dash although this is
a getter for a property that is set only in case we are not in dummy mode,
and so if the overview actor is defined.

So, to simplify the management, replace the overview's 'dash' property with
a dynamic getter that just returns the default value or our main dash
depending on the overview mode.
As per this we also need to toggle d2d in case the session mode changes.

Remove references to Main.overview._dash as in the current codebase is just
a dead-property.

Finally, always restore stock dash before destroying our dash to avoid we
could ever access to a dead actor.

Fixes LP: #1795703, LP: #1865865
Fixes #1121