docking: Reimplement Clutter.BindConstraint to bind to Maximum size only
With mutter commit 4f8e518d4 BindConstraint is applied in a way that the
source sizes are always applied to the bound widget, no matter if we need
to constrain it or not.

So reimplement Clutter.BindConstraint in the dock, although this needs a fix
in gjs that will be available only starting from gjs 1.65.2, so we need to
check for that first.

See more at:
 https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/1070