docking: Make sure we don't send the overview offscreen in extended mode
When we vertically extend the dash, due to the fact that the dash is part of
the overviewControls, we make its layoutManager to wrongly compute the
viewSelector size, that ends up offscreen.

So, workaround this by connecting to the overviewControl 'allocation-changed'
signal and compute the offset we need to apply not to go offscreen and we
set is as bottom margin.

Unfortunately, it looks like that even replacing the layout-manager with
another implementation to the overviewControls that does the same isn't
enough to make this not to happen, and so far I've not been able to track
down how to avoid this in any other way.

Fixes #1007
Fixes LP: #1834967