Get dash max size from the container.
This is following the upstram changes to the bindconstraints, and our
initial attempt to revert to the previous behaviour (see
2882d6bb7d4d31fe28006a6456f1d3e61ab8f617).

Unfortunately, the icon size machinery is still broken as the actual
size of the dash rather than the maximum one is used (I suspect becasue
the constrain machinery is not setting the preferredSize of the actor
and not just its allocation).

The easier solution is to get the size from one of the dash container,
and connect watch events there. Therefore make the maximum dash size a
parameter which is set from outside.

Fix #1169