Revert "Fix #1232".
Too many behaviour changes in the reverted commit (as correctly pointed
out by the author 4b165321f2b4633518a0261625e5613b22416ad8):

1) The hide animation is now always delayed, making it not in sync with
the rest of the animations (e.g. exiting the overview)

2) Not only the dock can be shown hitting any area of the screen border,
but the only way to hide is is to leave the border by the docker height
even after habing entered and interacted with the dock: this is not
necessary a bad behaviour, but it is un unintentional one.

This reverts commit a62b48bd7852a34064d539c71823b342afc119fd.