Fix #1232 again.
The problem originates from the transition to clutter animations:

662604b33f89f965e02664b63d1603be270e5c65

and the solution for a regression due to how "queued" transitions are handled
71abe800b2bc0bca60c63c8e646db2a8a9827ddf

The solution seems to correctly removed any "delayed" hide transition.
It was not canceled every time a dock show is requested, but only when
a new show transition was initiated (therefore if the show is requested
while the dock is already showing, the queued hide was not removed).